Move and rename personApi.ts
Same as the others
