+++ src/web/api/person/mutations/upsertPerson.ts
@@ -1,63 +1,19 @@
-import {
-    MutationUpsertOnePersonArgs,
-    Person,
-    PersonUserIdNameCompoundUniqueInput,
-} from "@client-types"
+import { MutationUpsertOnePersonArgs } from "@client-types"
 import {
     MutateFunction,
     UseMutationReturn,
-    UseQueryReturn,
     useMutation,
-    useQuery,
 } from "@vue/apollo-composable"
 import { gql } from "graphql-tag"
 
-import { createUserIdNameCompoundUniqueInputForUpdateOrUpsertOne } from "@/api/globalApiHelper"
+import { createUserIdNameCompoundUniqueInputForUpdateOrUpsertOne } from "@/api/globalApiHelper.js"
 import {
     createTagToObjectRelationCreateOrConnectInputForMutations,
     createTagToObjectRelationWhereUniqueInputForMutations,
 } from "@/api/tagToObjectRelation/tagToObjectRelationApiHelper.js"
-import { createUserWhereUniqueInputForMutations } from "@/api/user/userApiHelper"
+import { createUserWhereUniqueInputForMutations } from "@/api/user/userApiHelper.js"
 
-interface QueryGetPersonResult {
-    person: Person,
-}
 
-// I'm not sure why this is necessary. The type is essentially identical to QueryGetPersonArgs from @client-types
-interface CustomQueryGetPersonArgs {
-    where: {
-        userId_name: PersonUserIdNameCompoundUniqueInput
-    }
-}
-
-const personQueryText = gql`
-    query person($where: PersonWhereUniqueInput!) {
-        person(where: $where) {
-            id
-            name
-            description
-            tags {
-                tag {
-                    id
-                    name
-                }
-            }
-        }
-    }
-`
-export function doPersonQuery(
-    name: string,
-    userId: string,
-): UseQueryReturn<QueryGetPersonResult, CustomQueryGetPersonArgs> {
-    return useQuery(personQueryText, {
-        where: {
-            userId_name: {
-                name,
-                userId,
-            },
-        },
-    })
-}
 interface UpsertPersonResult {
     upsertOnePerson: {
         id: string
